requirements.txt
boto3
elevenlabs
flask
ipython
moviepy
nltk
numpy
opencv-python
openai
Pillow
python-dotenv
PyQt5
requests
scikit-learn
torch
ultralytics
colour-science

